fix: modifica valores monetarios de Double para Bigdecimal para evitar problemas de precisao

+++ backend/src/main/java/br/com/inthurn/backend/model/transport/request/CashBalanceRequestDTO.java
@@ -1,4 +1,6 @@
 package br.com.inthurn.backend.model.transport.request;
 
-public record CashBalanceRequestDTO(String id, String description, Double initialBalance) {
+import java.math.BigDecimal;
+
+public record CashBalanceRequestDTO(String id, String description, BigDecimal initialBalance) {
 }
